% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_matching_covariate}
\alias{add_matching_covariate}
\title{Add a Covariate to the Matching Set}
\usage{
add_matching_covariate(matching_information, ipd, covariate_name,
  target_mean = NULL, target_sd = NULL, target_proportion = NULL)
}
\description{
\code{add_matching_covariate} adds a covariate to the matching set initialised
by \code{set_matching}.
}
\details{
When the covariate is added the patient information will be added to the
\code{matching_set} object, the target will be added to $targets and the
relevant summary statistic of the covariate prior to matching will be
calculated and appended to the \code{matching_set}.
}
