% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic}
\alias{maic}
\title{Performing a Matching-Adjusted Indirect Comparison}
\usage{
maic(ipd, matching_covariates, outcome_identifier, arm_identifier,
  outcome_type = NULL, binary_event_type = NULL,
  comparator_to_anchor = NULL, se_comparator_to_anchor = NULL,
  name_intervention = "Intervention", name_anchor = "Anchor",
  name_comparator = "Comparator")
}
\description{
\code{maic} is used to perform a Matching_Adjusted Indirect Comparison. Given a set of individual patient data (IPD), a set of covariates to match on, \code{maic} will generate weights to be applied to the IPD in order to perform a Matching-Adjusted Indirect Comparison. Currently the package can match continuous covariates to a mean and standard deviation as well as binary covariates to a target proportion.
}
\details{
It is important that the IPD is in a particular format.
}
\section{IPD Format}{

A patient ID column is optional. The maicer package anticipates that IPD will be in a data frame format with each covariate listed in a separate column. For an anchored MAIC, a variable denoting the arm to which the patient was assigned is necessary and should follow the patient characteristic columns. The final column should give the outcome, y.
}

