% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maicer.R
\docType{package}
\name{maicer}
\alias{maicer}
\alias{maicer-package}
\title{maicer: Inverse Propensity Score Weighting for a Matching-Adjusted Indirect Comparison (MAIC)}
\description{
maicer: Inverse Propensity Score Weighting for a Matching-Adjusted Indirect Comparison (MAIC)
}
\section{Details}{

This package contains functions for performing a Matching-Adjusted Indirect Comparison.

For a complete list of functions, use \code{library(help = "maicer")}.
}

