% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_adjustment.R
\name{match_adjustment}
\alias{match_adjustment}
\title{Matching-Adjustment}
\usage{
match_adjustment(matching_object, outcomes = NULL,
  arm_identifier = NULL)
}
\description{
\code{match_adjustment} is used to estimate the weights needed to make the necessary adjustment for MAIC.
}
\details{
Specifying the covariates to match on begins with creating a "matching set"
which is made up of the individual level patient data and the summary
information of that covariate. This requires the type of the summary
statistic from the aggregate data to which the covariate needs to match to.
}
