% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matched_check}
\alias{matched_check}
\title{Verify matching of weighted IPD to target AgD.}
\usage{
matched_check(original_data, matched_data, targets, weights, ...)
}
\description{
\code{matched_check}
}
