% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_matching}
\alias{set_matching}
\title{Initialise the Set of Matching Covariates}
\usage{
set_matching(ipd, covariate_name, target_mean = NULL, target_sd = NULL,
  target_proportion = NULL)
}
\description{
\code{set_matching} creates a data frame of one covariate from the individual
patient data to be used for matching purposes and to calculate the weights.
}
\details{
Specifying the covariates to match on begins with creating a "matching set"
which is made up of the individual level patient data and the summary
information of that covariate. This requires the type of the summary
statistic from the aggregate data to which the covariate needs to match to.
}
